+++ src/templates/rainer.js
@@ -5,6 +5,7 @@ const bgImage = '//pages.convertkit.com/assets/rainier/bg.jpg'
 
 function Rainer({
   options,
+  className,
   action,
   submitText,
   emailPlaceholder,
@@ -19,7 +20,7 @@ function Rainer({
   stacked,
 }) {
   return (
-    <form action={action} className="seva-form formkit-form" method="post" target={newTab ? '_blank' : '_self'} data-sv-form={formId} data-uid={options.uid} data-format={format} data-version={options.version} data-options={JSON.stringify(options)} min-width="400 500 600 700 800">
+    <form action={action} className={`seva-form formkit-form rainer ${className}`} method="post" target={newTab ? '_blank' : '_self'} data-sv-form={formId} data-uid={options.uid} data-format={format} data-version={options.version} data-options={JSON.stringify(options)} min-width="400 500 600 700 800">
       <div className="formkit-container" data-style="full" style={{ backgroundColor: 'rgb(255, 255, 255)' }}>
         <div className="formkit-hero">
           <div className="formkit-hero-bg-color" style={{ background: `linear-gradient(rgba(16, 16, 16, .8), rgba(16, 16, 16, .8)), url(${bgImage})`, backgroundSize: 'cover', backgroundPosition: 'center' }}></div>